TST: split up test_concat.py #37243
 * created test_invalid.py

 * created test_sort.py
